REM: Initial changes to make it work (seemingly) with MinGW-w64.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@42601 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/recompiler/cpu-common.h
@@ -21,7 +21,7 @@
 #if !defined(CONFIG_USER_ONLY)
 
 /* address in the RAM (different from a physical address) */
-typedef unsigned long ram_addr_t;
+typedef uintptr_t ram_addr_t;
 
 /* memory API */
