+++ src/recompiler/softmmu_header.h
@@ -86,7 +86,7 @@ static inline RES_TYPE glue(glue(ld, USUFFIX), MEMSUFFIX)(target_ulong ptr)
     int page_index;
     RES_TYPE res;
     target_ulong addr;
-    unsigned long physaddr;
+    uintptr_t physaddr;
     int mmu_idx;
 
     addr = ptr;
@@ -107,7 +107,7 @@ static inline int glue(glue(lds, SUFFIX), MEMSUFFIX)(target_ulong ptr)
 {
     int res, page_index;
     target_ulong addr;
-    unsigned long physaddr;
+    uintptr_t physaddr;
     int mmu_idx;
 
     addr = ptr;
@@ -132,7 +132,7 @@ static inline void glue(glue(st, SUFFIX), MEMSUFFIX)(target_ulong ptr, RES_TYPE 
 {
     int page_index;
     target_ulong addr;
-    unsigned long physaddr;
+    uintptr_t physaddr;
     int mmu_idx;
 
     addr = ptr;
